Stop setting INFO_LOGONNOTIFY and INFO_LOGONERRORS
These flags are sent from the client to the RDP server during connection
establishment, and are used to request that the server notify the client
when the logon completes or errors.

These notifications then come back from the server in the form of a
"Save Session Info PDU" - which this library doesn't currently support.
This results in an error message being logged:

    GLOBAL: Parsing data PDU error RdpError(RdpError { kind: NotImplemented, message: "GLOBAL: Data PDU parsing not implemented Pdutype2SaveSessionInfo" })

Since we don't parse these notifications, there's no need to request
that the server send them.