Fix a potential panic when the server requests NLA
If the server attempts to start NLA authentication but the client
fails to provide the correct implementation, the code will panic
when unwrapping an empty option.

Return an error in this case instead of blindly unwrapping.

Note: this is unlikely to happen in practice unless the server is
misbehaving, because the client never requests ProtocolHybrid so
the server should not confirm it.