Clean up 32bpp RLE decoder
The existing code for this was a direct port of C code.
As a result, it had a bunch of variables with scope
larger than necessary and no documentation.

No functional changes here, just simplify the code and
add comments and links to the relevant parts of the spec
to make things easier to understand.

Included a new test to verify that behavior is the same
before and after the change.